Final Check-In

+++ URHealth.Service/BMI/BMICategory.cs
@@ -8,17 +8,17 @@ namespace URHealth.Service.BMI
 {
     public static class BMICategory
     {
-        public static string GetBMICategory(decimal bmiValue)
+        public static string GetBMICategory(float bmiValue)
         {
             string bmiCategory = "";
 
-            if (bmiValue <= (decimal)18.5)
+            if (bmiValue <= 18.5f)
                 bmiCategory = "Under Weight";
 
-            else if (bmiValue > (decimal)18.5 && bmiValue <= (decimal)24.9)
+            else if (bmiValue > 18.5f && bmiValue <= 24.9f)
                 bmiCategory = "Normal Weight";
 
-            else if (bmiValue > (decimal)25 && bmiValue <= (decimal)29.9)
+            else if (bmiValue > 25f && bmiValue <= 29.9f)
                 bmiCategory = "Over Weight";
             else
                 bmiCategory = "Obesity";
